[IMP] mail, *: tests: use mail.channel.partner for seen indicator
* = hr_holidays, im_livechat, website_livechat

Task-2810539

closes odoo/odoo#87587

Related: odoo/enterprise#25911

+++ addons/mail/static/src/models/channel_invitation_form.js
@@ -2,7 +2,7 @@
 
 import { registerModel } from '@mail/model/model_core';
 import { attr, many, one } from '@mail/model/model_field';
-import { clear, insertAndReplace, link, replace, unlink, unlinkAll } from '@mail/model/model_field_command';
+import { clear, insertAndReplace, link, replace, unlink } from '@mail/model/model_field_command';
 import { cleanSearchTerm } from '@mail/utils/utils';
 
 registerModel({
@@ -57,11 +57,9 @@ registerModel({
                     },
                 }));
             }
-            this.update({
-                searchTerm: "",
-                selectedPartners: unlinkAll(),
-            });
-            this.delete();
+            if (this.exists()) {
+                this.delete();
+            }
         },
         /**
          * @param {Partner} partner
